workflow/requirements.txt
apache-airflow-providers-apache-kafka==1.6.0
apache-airflow-providers-amazon==8.28.0
apache-airflow-providers-google==10.22.0
apache-airflow-providers-snowflake==5.7.0
apache-airflow-providers-slack==8.9.0
apache-airflow-providers-mysql==5.7.0
apache-airflow-providers-redis==3.8.0
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-apache-spark==4.10.0
numpy==1.26.4
pandas==2.1.4
boto3==1.35.16
openpyxl==3.1.5
slack_sdk==3.32.0
kafka-python-ng==2.2.3
shortuuid==1.0.13
jupyterlab